split dense buffer into separate file
